backend/requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
decorator==5.1.0
Django==3.2.8
django-filter==21.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==3.3
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
numpy==1.21.3
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.21
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
traitlets==5.1.1
urllib3==1.26.7
uWSGI==2.0.20
wcwidth==0.2.5
